// js-storefront/tbsStore/src/app/interfaces/custom-search-box.interface.ts
export interface SearchBoxConfig {
  displaySuggestions: boolean;
  displayProducts: boolean;
  displayProductImages?: boolean;
  maxProducts?: number;
  maxSuggestions?: number;
  minCharactersBeforeRequest?: number;
}
export interface SearchResults {
  message?: string;
  products?: any[];
  suggestions?: string[];
  keywordRedirectUrl?: string;
}